LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/blob-popup.https-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'window.opener.furtherPopup = w')

FAIL Cross-Origin-Opener-Policy and a blob URL popup assert_equals: expected 0 but got 36
